requirement.txt
apturl==0.5.2
attrs==19.3.0
Automat==0.8.0
backports.entry-points-selectable==1.1.1
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.0.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crypto==1.4.1
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
deluge==2.0.3
distlib==0.3.3
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
evdev==1.3.0
filelock==3.4.0
Flask==2.0.2
Flask-HTTPAuth==4.5.0
GeoIP==1.3.2
Guake==3.6.3
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
incremental==16.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
meson==0.53.2
Naked==0.1.31
netifaces==0.10.4
networkx==2.6.3
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.4.60
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.4.0
protobuf==3.6.1
psutil==5.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygame==1.9.6
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-libtorrent==1.1.13
python-magic==0.4.16
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
rencode==1.0.6
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
shellescape==3.8.1
simplejson==3.16.0
six==1.14.0
soupsieve==1.9.5
systemd-python==234
torch==1.10.0
Twisted==18.9.0
typing-extensions==4.0.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.10.0
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==2.0.2
xkit==0.0.0
zope.interface==4.7.1
